requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
ffmpeg-python==0.2.0
filelock==3.9.0
frozenlist==1.3.3
future==0.18.3
huggingface-hub==0.12.1
idna==3.4
more-itertools==9.1.0
multidict==6.0.4
numpy==1.24.2
openai==0.27.0
openai-whisper==20230124
packaging==23.0
PyAudio==0.2.13
pydub==0.25.1
python-dotenv==1.0.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
semantic-version==2.10.0
setuptools-rust==1.5.2
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
